+++ src/tooling/piral-cli/src/apps/publish-piral.ts
@@ -1,7 +1,17 @@
 import { resolve } from 'path';
 import { publishArtifacts } from '../release';
 import { LogLevels, PiralBuildType } from '../types';
-import { setLogLevel, progress, checkExists, fail, logDone, logReset, publishNpmPackage, matchFiles } from '../common';
+import {
+  setLogLevel,
+  progress,
+  checkExists,
+  fail,
+  logDone,
+  logReset,
+  publishNpmPackage,
+  matchFiles,
+  log,
+} from '../common';
 
 export interface PublishPiralOptions {
   /**
@@ -15,9 +25,14 @@ export interface PublishPiralOptions {
   logLevel?: LogLevels;
 
   /**
-   * The additional fields to supply for the provider.
+   * The options to supply for the provider.
    */
-  fields?: Record<string, string>;
+  opts?: Record<string, string>;
+
+  /**
+   * Defines if authorization tokens can be retrieved interactively.
+   */
+  interactive?: boolean;
 
   /**
    * The provider to use for publishing the release artifacts.
@@ -35,10 +50,16 @@ export const publishPiralDefaults: PublishPiralOptions = {
   logLevel: LogLevels.info,
   type: 'all',
   provider: 'none',
-  fields: {},
+  opts: {},
+  interactive: false,
 };
 
-async function publishEmulator(baseDir: string, source: string, args: Record<string, string> = {}) {
+async function publishEmulator(
+  baseDir: string,
+  source: string,
+  args: Record<string, string> = {},
+  interactive = false,
+) {
   const type = 'emulator';
   const directory = resolve(baseDir, source, type);
   const exists = await checkExists(directory);
@@ -58,7 +79,7 @@ async function publishEmulator(baseDir: string, source: string, args: Record<str
     p.push(`--${c}`, args[c]);
     return p;
   }, [] as Array<string>);
-  await publishNpmPackage(directory, file, flags);
+  await publishNpmPackage(directory, file, flags, interactive);
 }
 
 async function publishRelease(
@@ -66,6 +87,7 @@ async function publishRelease(
   source: string,
   providerName: string,
   args: Record<string, string> = {},
+  interactive = false,
 ) {
   const type = 'release';
   const directory = resolve(baseDir, source, type);
@@ -76,7 +98,8 @@ async function publishRelease(
   }
 
   const files = await matchFiles(directory, '**/*');
-  await publishArtifacts(providerName, files, args);
+  log('generalDebug_0003', `Found ${files.length} in "${directory}": ${files.join(', ')}`);
+  await publishArtifacts(providerName, directory, files, args, interactive);
 }
 
 export async function publishPiral(baseDir = process.cwd(), options: PublishPiralOptions = {}) {
@@ -84,8 +107,9 @@ export async function publishPiral(baseDir = process.cwd(), options: PublishPira
     source = publishPiralDefaults.source,
     type = publishPiralDefaults.type,
     logLevel = publishPiralDefaults.logLevel,
-    fields = publishPiralDefaults.fields,
+    opts = publishPiralDefaults.opts,
     provider = publishPiralDefaults.provider,
+    interactive = publishPiralDefaults.interactive,
   } = options;
   const fullBase = resolve(process.cwd(), baseDir);
   setLogLevel(logLevel);
@@ -98,14 +122,14 @@ export async function publishPiral(baseDir = process.cwd(), options: PublishPira
 
   if (type !== 'release') {
     progress('Publishing emulator package ...');
-    await publishEmulator(fullBase, source, fields);
+    await publishEmulator(fullBase, source, opts, interactive);
     logDone(`Successfully published emulator.`);
     logReset();
   }
 
   if (type !== 'emulator') {
     progress('Publishing release files ...');
-    await publishRelease(fullBase, source, provider, fields);
+    await publishRelease(fullBase, source, provider, opts, interactive);
     logDone(`Successfully published release.`);
     logReset();
   }